Created network aware mixin
