+++ src/client/mixins/index.ts
@@ -5,3 +5,4 @@ export * from './marketo';
 export * from './google-analytics';
 export * from './lighthouse';
 export * from './linked-in';
+export * from './network';
